Restructure component directory
