+++ app/react/src/reactRouter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Section2b from "./components/sections/Section2B";
-import Section3c from "./components/sections/Section3c";
+import Section2b from "./components/sections/section2b/Section2B";
+import Section3c from "./components/sections/section3c/Section3c";
 import Homepage from "./components/Homepage";
 import Preamble from "./components/sections/Preamble";
 import BasicInfo from "./components/sections/BasicInfo";
